api/requirements.txt
flask
flask-cors
requests
python-dotenv
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
pymysql
cryptography
